pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "animatr"
version = "0.1.0"
description = "Motor declarativo de animación audiovisual"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "ANIMATR Team" }
]
keywords = ["animation", "video", "yaml", "moho", "blender", "tts"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "rich>=13.0",
    "httpx>=0.25",
    "openai>=1.0",
    "mutagen>=1.47",
    "requests>=2.31",
    # AI Agents dependencies
    "crewai>=0.86",
    "crewai-tools>=0.14",
    "langchain-openai>=0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
]

[project.scripts]
animatr = "animatr.cli:app"

[project.urls]
Homepage = "https://github.com/nadalpiantini/animatr"
Documentation = "https://github.com/nadalpiantini/animatr#readme"
Repository = "https://github.com/nadalpiantini/animatr"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
